Move test helper functions to separate file
