+++ rust-src/wallet_library/src/lib.rs
@@ -1,2 +1,4 @@
 pub mod identity;
+#[cfg(test)]
+mod test_helpers;
 pub mod wallet;
